+++ tests/testthat/test-cleaning-utils.R
@@ -6,34 +6,37 @@ test$words_factor <- as.factor(test$words)
 test$numbers_characters <- as.character(test$numbers)
 
 test_that("Pad decimals works", {
-  expect_equal(pad_decimals(2.11, 1), 2.1)
-  expect_equal(pad_decimals(2.11, 0), 2)
+  expect_equal(pad_decimals(2.11, 1), "2.1")
+  expect_equal(pad_decimals(2.11, 0), "2")
 
-  expect_equal(pad_decimals(2, 2), 2.00)
-  expect_equal(pad_decimals(2, 1), 2.0)
-  expect_equal(pad_decimals(2, 5), 2.00000)
-  expect_equal(pad_decimals(c(2, 3.11), 2), c(2.00, 3.11))
+  expect_equal(pad_decimals(2, 2), "2.00")
+  expect_equal(pad_decimals(2, 1), "2.0")
+  expect_equal(pad_decimals(2, 5), "2.00000")
+  expect_equal(pad_decimals(c(2, 3.11), 2), c("2.00", "3.11"))
+  expect_equal(pad_decimals(c(2, 3.11), 3), c("2.000", "3.110"))
 
 
-  expect_equal(pad_decimals(2.11, 2), 2.11)
-  expect_equal(pad_decimals(2.11, 3), 2.110)
-  expect_equal(pad_decimals(2.11, 7), 2.1100000)
+  expect_equal(pad_decimals(2.11, 2), "2.11")
+  expect_equal(pad_decimals(c(2.11, NA), 2), c("2.11", NA))
+  expect_equal(pad_decimals(2.11, 3), "2.110")
+  expect_equal(pad_decimals(2.11, 7), "2.1100000")
   expect_error(pad_decimals("2.11", 1))
-  expect_equal(pad_decimals(c(2.3, 8, 6.789)), c(2.300, 8.000, 6.789))
+  expect_equal(pad_decimals(c(2.3, 8, 6.789)), c("2.300", "8.000", "6.789"))
   expect_error(pad_decimals(as.factor(c(2.3, 8, 6.789))))
   expect_error(pad_decimals(c("2.3", "8", "6.789")))
-  expect_equal(pad_decimals(c(2.3, 8, 6.789), 0), c(2, 8, 7))
+  expect_equal(pad_decimals(c(2.3, 8, 6.789), 0), c("2", "8", "7"))
 
-  expect_equal(pad_decimals(c(2.3, 8.5, 8.4, 7.5, 6.789), 0), c(2, 8, 8, 8, 7))
+  expect_equal(pad_decimals(c(2.3, 8.5, 8.4, 7.5, 6.789), 0), c("2", "8", "8", "8", "7"))
   expect_equal(pad_decimals(c(2.3, 8.5, 8.4, 7.5, 6.789), 2),
-               c(2.30, 8.50, 8.40, 7.50, 6.79))
+               c("2.30", "8.50", "8.40", "7.50", "6.79"))
 
 
 
-  expect_type(pad_decimals(2.11, 2), "double")
-  expect_type(pad_decimals(2.11, 3), "double")
-  expect_type(pad_decimals(2.11, 7), "double")
-  expect_type(pad_decimals(c(2.3, 8, 6.789)), "double")
+  expect_type(pad_decimals(2.11, 2), "character")
+  expect_type(pad_decimals(2.11, 3), "character")
+  expect_type(pad_decimals(2.11, 7), "character")
+  expect_type(pad_decimals(c(2.11, NA), 7), "character")
+  expect_type(pad_decimals(c(2.3, 8, 6.789)), "character")
 
 
 })