+++ NEWS.md
@@ -1,7 +1,9 @@
 # crime 0.2.0
 
 * Adds a function to that returns a data.frame from a regression model
-so it's easier to turn into tables for publication.
+so it's easier to turn into tables for publication. Also adds a function
+that graphs the coefficient value and the 95% confidence interval
+for all coefficients.
 
 # crime 0.1.0