+++ NEWS.md
@@ -4,6 +4,9 @@
 so it's easier to turn into tables for publication. Also adds a function
 that graphs the coefficient value and the 95% confidence interval
 for all coefficients.
+* Updates the LaTeX table generator code to improve formatting.
+* Fixed bug in `pad_decimals()` that incorrectly dropped the padding when converting back to a number. Now
+returns data as strings.
 
 # crime 0.1.0